Osa 4 suoritus

+++ osa4/controllers/blogs.js
@@ -1,35 +1,51 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog.js")
+const middleware = require("../utils/middleware")
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 })
     response.json(blogs)
 })
 
-blogsRouter.get('/', async (request, response) => {
-    const blog = await Blog.findById(request.params.id)
+blogsRouter.get('/:id', async (request, response) => {
+    const blog = await Blog.findById(request.params.id).populate("user", { username: 1, name: 1 })
     if (blog) {
-        response.status(200).json(blog)
+        response.json(blog.toJSON())
     }
     else {
         response.status(400)
     }
 })
 
-blogsRouter.post('/', async (request, response) => {
-    if (!("title" in request.body) || !("url" in request.body)) {
-        response.status(400).end("title or url missing")
-        return;
-    }
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+    const body = request.body
+    const user = request.user
+    const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        user: user._id,
+        likes: body.likes || 0
+    })
 
-    const blog = new Blog(request.body)
-    blog.likes = blog.likes || 0
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
 
-    const result = await blog.save()
-    response.status(200).json(result)
+    response.json(savedBlog.toJSON())
 })
 
-blogsRouter.delete("/:id", async (request, response) => {
+blogsRouter.delete("/:id", middleware.userExtractor, async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).end()
+    }
+
+    if (user._id.toString() !== blog.user.toString()) {
+        return response.status(401).json({ error: "You are not authorized to delete this blog" })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(200).end()
 })
@@ -41,10 +57,11 @@ blogsRouter.put("/:id", async (request, response) => {
         title: request.body.title,
         author: request.body.author,
         url: request.body.url,
+        user: request.body.user,
         likes: request.body.likes
     }
 
-    const updatedNote = await Blog.findByIdAndUpdate(id, blog, { new: true })
+    const updatedNote = await Blog.findByIdAndUpdate(id, blog, { new: true }).populate("user")
     response.json(updatedNote.toJSON())
 })
